:recycle: increments to bigIncrements for users table

+++ database/migrations/1708549951032_create_access_tokens_table.ts
@@ -5,9 +5,9 @@ export default class extends BaseSchema {
 
   async up() {
     this.schema.createTable(this.tableName, (table) => {
-      table.increments('id')
+      table.bigIncrements('id')
       table
-        .integer('tokenable_id')
+        .bigInteger('tokenable_id')
         .notNullable()
         .unsigned()
         .references('id')
