+++ database/migrations/1710328631042_create_testimonials_table.ts
@@ -6,7 +6,7 @@ export default class extends BaseSchema {
   async up() {
     this.schema.createTable(this.tableName, (table) => {
       table.increments('id')
-      table.integer('author_id').unsigned().references('id').inTable('users')
+      table.bigInteger('author_id').unsigned().references('id').inTable('users')
       table.integer('project_id').unsigned().references('id').inTable('projects').nullable()
       table.text('content').notNullable()
       table.integer('stars').notNullable()
