+++ database/migrations/1713545464917_create_saves_table.ts
@@ -6,7 +6,7 @@ export default class extends BaseSchema {
   async up() {
     this.schema.createTable(this.tableName, (table) => {
       table.increments('id')
-      table.integer('author_id').notNullable()
+      table.bigInteger('author_id').notNullable()
       table.string('name').notNullable()
       table.text('description').notNullable()
       table.jsonb('data').notNullable()
